chore: consolidate types
